requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pydantic-settings
greenlet
fastapi-cli

greenlet
